Refactoring entities to entity
