+++ apps/survey/src/entity/survey.entity.ts
